remove outer join from series trait
