+++ polars/polars-core/src/frame/hash_join/mod.rs
@@ -1,4 +1,6 @@
 pub(crate) mod multiple_keys;
+mod single_keys;
+mod single_keys_dispatch;
 
 use polars_arrow::utils::CustomIterTools;
 
@@ -23,73 +25,10 @@ use std::hash::{BuildHasher, Hash, Hasher};
 pub use self::multiple_keys::private_left_join_multiple_keys;
 use crate::frame::groupby::hashing::HASHMAP_INIT_SIZE;
 use crate::utils::series::to_physical_and_bit_repr;
-
-/// If Categorical types are created without a global string cache or under
-/// a different global string cache the mapping will be incorrect.
-#[cfg(feature = "dtype-categorical")]
-pub(crate) fn check_categorical_src(l: &DataType, r: &DataType) -> Result<()> {
-    match (l, r) {
-        (DataType::Categorical(Some(l)), DataType::Categorical(Some(r))) => {
-            if !l.same_src(&*r) {
-                return Err(PolarsError::ComputeError("joins/or comparisons on categorical dtypes can only happen if they are created under the same global string cache".into()));
-            }
-            Ok(())
-        }
-        _ => Ok(()),
-    }
-}
-
-impl Series {
-    #[cfg(feature = "private")]
-    #[doc(hidden)]
-    pub fn hash_join_left(&self, other: &Series) -> Vec<(IdxSize, Option<IdxSize>)> {
-        let (lhs, rhs) = (self.to_physical_repr(), other.to_physical_repr());
-
-        use DataType::*;
-        match lhs.dtype() {
-            Utf8 => {
-                let lhs = lhs.utf8().unwrap();
-                let rhs = rhs.utf8().unwrap();
-                lhs.hash_join_left(rhs)
-            }
-            _ => {
-                if self.bit_repr_is_large() {
-                    let lhs = self.bit_repr_large();
-                    let rhs = other.bit_repr_large();
-                    num_group_join_left(&lhs, &rhs)
-                } else {
-                    let lhs = self.bit_repr_small();
-                    let rhs = other.bit_repr_small();
-                    num_group_join_left(&lhs, &rhs)
-                }
-            }
-        }
-    }
-
-    fn hash_join_inner(&self, other: &Series) -> Vec<(IdxSize, IdxSize)> {
-        let (lhs, rhs) = (self.to_physical_repr(), other.to_physical_repr());
-
-        use DataType::*;
-        match lhs.dtype() {
-            Utf8 => {
-                let lhs = lhs.utf8().unwrap();
-                let rhs = rhs.utf8().unwrap();
-                lhs.hash_join_inner(rhs)
-            }
-            _ => {
-                if self.bit_repr_is_large() {
-                    let lhs = self.bit_repr_large();
-                    let rhs = other.bit_repr_large();
-                    num_group_join_inner(&lhs, &rhs)
-                } else {
-                    let lhs = self.bit_repr_small();
-                    let rhs = other.bit_repr_small();
-                    num_group_join_inner(&lhs, &rhs)
-                }
-            }
-        }
-    }
-}
+#[cfg(feature = "asof_join")]
+pub(crate) use single_keys::create_probe_table;
+#[cfg(feature = "asof_join")]
+pub(crate) use single_keys_dispatch::prepare_strs;
 
 macro_rules! det_hash_prone_order {
     ($self:expr, $other:expr) => {{
@@ -107,6 +46,23 @@ macro_rules! det_hash_prone_order {
 
         (a, b, !left_first)
     }};
+}
+
+pub(super) use det_hash_prone_order;
+
+/// If Categorical types are created without a global string cache or under
+/// a different global string cache the mapping will be incorrect.
+#[cfg(feature = "dtype-categorical")]
+pub(crate) fn check_categorical_src(l: &DataType, r: &DataType) -> Result<()> {
+    match (l, r) {
+        (DataType::Categorical(Some(l)), DataType::Categorical(Some(r))) => {
+            if !l.same_src(&*r) {
+                return Err(PolarsError::ComputeError("joins/or comparisons on categorical dtypes can only happen if they are created under the same global string cache".into()));
+            }
+            Ok(())
+        }
+        _ => Ok(()),
+    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -150,688 +106,6 @@ unsafe fn get_hash_tbl_threaded_join_mut_partitioned<T, H>(
         }
     }
     hash_tables.get_unchecked_mut(idx)
-}
-
-/// Probe the build table and add tuples to the results (inner join)
-fn probe_inner<T, F>(
-    probe: &[T],
-    hash_tbls: &[PlHashMap<T, Vec<IdxSize>>],
-    results: &mut Vec<(IdxSize, IdxSize)>,
-    local_offset: usize,
-    n_tables: u64,
-    swap_fn: F,
-) where
-    T: Send + Hash + Eq + Sync + Copy + AsU64,
-    F: Fn(IdxSize, IdxSize) -> (IdxSize, IdxSize),
-{
-    assert!(hash_tbls.len().is_power_of_two());
-    probe.iter().enumerate().for_each(|(idx_a, k)| {
-        let idx_a = (idx_a + local_offset) as IdxSize;
-        // probe table that contains the hashed value
-        let current_probe_table =
-            unsafe { get_hash_tbl_threaded_join_partitioned(k.as_u64(), hash_tbls, n_tables) };
-
-        let value = current_probe_table.get(k);
-
-        if let Some(indexes_b) = value {
-            let tuples = indexes_b.iter().map(|&idx_b| swap_fn(idx_a, idx_b));
-            results.extend(tuples);
-        }
-    });
-}
-
-pub(crate) fn create_probe_table<T, IntoSlice>(
-    keys: Vec<IntoSlice>,
-) -> Vec<PlHashMap<T, Vec<IdxSize>>>
-where
-    T: Send + Hash + Eq + Sync + Copy + AsU64,
-    IntoSlice: AsRef<[T]> + Send + Sync,
-{
-    let n_partitions = set_partition_size();
-
-    // We will create a hashtable in every thread.
-    // We use the hash to partition the keys to the matching hashtable.
-    // Every thread traverses all keys/hashes and ignores the ones that doesn't fall in that partition.
-    POOL.install(|| {
-        (0..n_partitions).into_par_iter().map(|partition_no| {
-            let partition_no = partition_no as u64;
-
-            let mut hash_tbl: PlHashMap<T, Vec<IdxSize>> =
-                PlHashMap::with_capacity(HASHMAP_INIT_SIZE);
-
-            let n_partitions = n_partitions as u64;
-            let mut offset = 0;
-            for keys in &keys {
-                let keys = keys.as_ref();
-                let len = keys.len() as IdxSize;
-
-                let mut cnt = 0;
-                keys.iter().for_each(|k| {
-                    let idx = cnt + offset;
-                    cnt += 1;
-
-                    if this_partition(k.as_u64(), partition_no, n_partitions) {
-                        let entry = hash_tbl.entry(*k);
-
-                        match entry {
-                            Entry::Vacant(entry) => {
-                                entry.insert(vec![idx]);
-                            }
-                            Entry::Occupied(mut entry) => {
-                                let v = entry.get_mut();
-                                v.push(idx);
-                            }
-                        }
-                    }
-                });
-                offset += len;
-            }
-            hash_tbl
-        })
-    })
-    .collect()
-}
-
-fn hash_join_tuples_inner<T, IntoSlice>(
-    probe: Vec<IntoSlice>,
-    build: Vec<IntoSlice>,
-    // Because b should be the shorter relation we could need to swap to keep left left and right right.
-    swap: bool,
-) -> Vec<(IdxSize, IdxSize)>
-where
-    IntoSlice: AsRef<[T]> + Send + Sync,
-    T: Send + Hash + Eq + Sync + Copy + AsU64,
-{
-    // NOTE: see the left join for more elaborate comments
-
-    // first we hash one relation
-    let hash_tbls = create_probe_table(build);
-
-    let n_tables = hash_tbls.len() as u64;
-    debug_assert!(n_tables.is_power_of_two());
-    let offsets = probe
-        .iter()
-        .map(|ph| ph.as_ref().len())
-        .scan(0, |state, val| {
-            let out = *state;
-            *state += val;
-            Some(out)
-        })
-        .collect::<Vec<_>>();
-    // next we probe the other relation
-    // code duplication is because we want to only do the swap check once
-    POOL.install(|| {
-        probe
-            .into_par_iter()
-            .zip(offsets)
-            .map(|(probe, offset)| {
-                let probe = probe.as_ref();
-                // local reference
-                let hash_tbls = &hash_tbls;
-                let mut results = Vec::with_capacity(probe.len());
-                let local_offset = offset;
-
-                // branch is to hoist swap out of the inner loop.
-                if swap {
-                    probe_inner(
-                        probe,
-                        hash_tbls,
-                        &mut results,
-                        local_offset,
-                        n_tables,
-                        |idx_a, idx_b| (idx_b, idx_a),
-                    )
-                } else {
-                    probe_inner(
-                        probe,
-                        hash_tbls,
-                        &mut results,
-                        local_offset,
-                        n_tables,
-                        |idx_a, idx_b| (idx_a, idx_b),
-                    )
-                }
-
-                results
-            })
-            .flatten()
-            .collect()
-    })
-}
-
-fn hash_join_tuples_left<T, IntoSlice>(
-    probe: Vec<IntoSlice>,
-    build: Vec<IntoSlice>,
-) -> Vec<(IdxSize, Option<IdxSize>)>
-where
-    IntoSlice: AsRef<[T]> + Send + Sync,
-    T: Send + Hash + Eq + Sync + Copy + AsU64,
-{
-    // first we hash one relation
-    let hash_tbls = create_probe_table(build);
-
-    // we determine the offset so that we later know which index to store in the join tuples
-    let offsets = probe
-        .iter()
-        .map(|ph| ph.as_ref().len())
-        .scan(0, |state, val| {
-            let out = *state;
-            *state += val;
-            Some(out)
-        })
-        .collect::<Vec<_>>();
-
-    let n_tables = hash_tbls.len() as u64;
-    debug_assert!(n_tables.is_power_of_two());
-
-    // next we probe the other relation
-    POOL.install(|| {
-        probe
-            .into_par_iter()
-            .zip(offsets)
-            // probes_hashes: Vec<u64> processed by this thread
-            // offset: offset index
-            .map(|(probe, offset)| {
-                // local reference
-                let hash_tbls = &hash_tbls;
-                let probe = probe.as_ref();
-
-                // assume the result tuples equal lenght of the no. of hashes processed by this thread.
-                let mut results = Vec::with_capacity(probe.len());
-
-                probe.iter().enumerate().for_each(|(idx_a, k)| {
-                    let idx_a = (idx_a + offset) as IdxSize;
-                    // probe table that contains the hashed value
-                    let current_probe_table = unsafe {
-                        get_hash_tbl_threaded_join_partitioned(k.as_u64(), hash_tbls, n_tables)
-                    };
-
-                    // we already hashed, so we don't have to hash again.
-                    let value = current_probe_table.get(k);
-
-                    match value {
-                        // left and right matches
-                        Some(indexes_b) => {
-                            results.extend(indexes_b.iter().map(|&idx_b| (idx_a, Some(idx_b))))
-                        }
-                        // only left values, right = null
-                        None => results.push((idx_a, None)),
-                    }
-                });
-                results
-            })
-            .flatten()
-            .collect()
-    })
-}
-
-/// Probe the build table and add tuples to the results (inner join)
-fn probe_outer<T, F, G, H>(
-    probe_hashes: &[Vec<(u64, T)>],
-    hash_tbls: &mut [PlHashMap<T, (bool, Vec<IdxSize>)>],
-    results: &mut Vec<(Option<IdxSize>, Option<IdxSize>)>,
-    n_tables: u64,
-    // Function that get index_a, index_b when there is a match and pushes to result
-    swap_fn_match: F,
-    // Function that get index_a when there is no match and pushes to result
-    swap_fn_no_match: G,
-    // Function that get index_b from the build table that did not match any in A and pushes to result
-    swap_fn_drain: H,
-) where
-    T: Send + Hash + Eq + Sync + Copy,
-    // idx_a, idx_b -> ...
-    F: Fn(IdxSize, IdxSize) -> (Option<IdxSize>, Option<IdxSize>),
-    // idx_a -> ...
-    G: Fn(IdxSize) -> (Option<IdxSize>, Option<IdxSize>),
-    // idx_b -> ...
-    H: Fn(IdxSize) -> (Option<IdxSize>, Option<IdxSize>),
-{
-    // needed for the partition shift instead of modulo to make sense
-    assert!(n_tables.is_power_of_two());
-    let mut idx_a = 0;
-    for probe_hashes in probe_hashes {
-        for (h, key) in probe_hashes {
-            let h = *h;
-            // probe table that contains the hashed value
-            let current_probe_table =
-                unsafe { get_hash_tbl_threaded_join_mut_partitioned(h, hash_tbls, n_tables) };
-
-            let entry = current_probe_table
-                .raw_entry_mut()
-                .from_key_hashed_nocheck(h, key);
-
-            match entry {
-                // match and remove
-                RawEntryMut::Occupied(mut occupied) => {
-                    let (tracker, indexes_b) = occupied.get_mut();
-                    *tracker = true;
-                    results.extend(indexes_b.iter().map(|&idx_b| swap_fn_match(idx_a, idx_b)))
-                }
-                // no match
-                RawEntryMut::Vacant(_) => results.push(swap_fn_no_match(idx_a)),
-            }
-            idx_a += 1;
-        }
-    }
-
-    for hash_tbl in hash_tbls {
-        hash_tbl.iter().for_each(|(_k, (tracker, indexes_b))| {
-            // remaining joined values from the right table
-            if !*tracker {
-                results.extend(indexes_b.iter().map(|&idx_b| swap_fn_drain(idx_b)))
-            }
-        });
-    }
-}
-
-/// Hash join outer. Both left and right can have no match so Options
-fn hash_join_tuples_outer<T, I, J>(
-    a: Vec<I>,
-    b: Vec<J>,
-    swap: bool,
-) -> Vec<(Option<IdxSize>, Option<IdxSize>)>
-where
-    I: Iterator<Item = T> + Send + TrustedLen,
-    J: Iterator<Item = T> + Send + TrustedLen,
-    T: Hash + Eq + Copy + Sync + Send,
-{
-    // This function is partially multi-threaded.
-    // Parts that are done in parallel:
-    //  - creation of the probe tables
-    //  - creation of the hashes
-
-    // during the probe phase values are removed from the tables, that's done single threaded to
-    // keep it lock free.
-
-    let size = a.iter().map(|a| a.size_hint().0).sum::<usize>()
-        + b.iter().map(|b| b.size_hint().0).sum::<usize>();
-    let mut results = Vec::with_capacity(size);
-
-    // prepare hash table
-    let mut hash_tbls = prepare_hashed_relation_threaded(b);
-    let random_state = hash_tbls[0].hasher().clone();
-
-    // we pre hash the probing values
-    let (probe_hashes, _) = create_hash_and_keys_threaded_vectorized(a, Some(random_state));
-
-    let n_tables = hash_tbls.len() as u64;
-
-    // probe the hash table.
-    // Note: indexes from b that are not matched will be None, Some(idx_b)
-    // Therefore we remove the matches and the remaining will be joined from the right
-
-    // branch is because we want to only do the swap check once
-    if swap {
-        probe_outer(
-            &probe_hashes,
-            &mut hash_tbls,
-            &mut results,
-            n_tables,
-            |idx_a, idx_b| (Some(idx_b), Some(idx_a)),
-            |idx_a| (None, Some(idx_a)),
-            |idx_b| (Some(idx_b), None),
-        )
-    } else {
-        probe_outer(
-            &probe_hashes,
-            &mut hash_tbls,
-            &mut results,
-            n_tables,
-            |idx_a, idx_b| (Some(idx_a), Some(idx_b)),
-            |idx_a| (Some(idx_a), None),
-            |idx_b| (None, Some(idx_b)),
-        )
-    }
-    results
-}
-
-pub(crate) trait HashJoin<T> {
-    fn hash_join_inner(&self, _other: &ChunkedArray<T>) -> Vec<(IdxSize, IdxSize)> {
-        unimplemented!()
-    }
-    fn hash_join_left(&self, _other: &ChunkedArray<T>) -> Vec<(IdxSize, Option<IdxSize>)> {
-        unimplemented!()
-    }
-    fn hash_join_outer(&self, _other: &ChunkedArray<T>) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        unimplemented!()
-    }
-}
-
-impl HashJoin<Float32Type> for Float32Chunked {
-    fn hash_join_outer(&self, other: &Float32Chunked) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        let ca = self.bit_repr_small();
-        let other = other.bit_repr_small();
-        ca.hash_join_outer(&other)
-    }
-}
-
-impl HashJoin<Float64Type> for Float64Chunked {
-    fn hash_join_outer(&self, other: &Float64Chunked) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        let ca = self.bit_repr_large();
-        let other = other.bit_repr_large();
-        ca.hash_join_outer(&other)
-    }
-}
-
-fn num_group_join_inner<T>(
-    left: &ChunkedArray<T>,
-    right: &ChunkedArray<T>,
-) -> Vec<(IdxSize, IdxSize)>
-where
-    T: PolarsIntegerType,
-    T::Native: Hash + Eq + Send + AsU64 + Copy,
-    Option<T::Native>: AsU64,
-{
-    let n_threads = POOL.current_num_threads();
-    let (a, b, swap) = det_hash_prone_order!(left, right);
-    let splitted_a = split_ca(a, n_threads).unwrap();
-    let splitted_b = split_ca(b, n_threads).unwrap();
-    match (
-        left.null_count() == 0,
-        right.null_count() == 0,
-        left.chunks.len(),
-        right.chunks.len(),
-    ) {
-        (true, true, 1, 1) => {
-            let keys_a = splitted_a
-                .iter()
-                .map(|ca| ca.cont_slice().unwrap())
-                .collect::<Vec<_>>();
-            let keys_b = splitted_b
-                .iter()
-                .map(|ca| ca.cont_slice().unwrap())
-                .collect::<Vec<_>>();
-            hash_join_tuples_inner(keys_a, keys_b, swap)
-        }
-        (true, true, _, _) => {
-            let keys_a = splitted_a
-                .iter()
-                .map(|ca| ca.into_no_null_iter().collect::<Vec<_>>())
-                .collect::<Vec<_>>();
-            let keys_b = splitted_b
-                .iter()
-                .map(|ca| ca.into_no_null_iter().collect::<Vec<_>>())
-                .collect::<Vec<_>>();
-            hash_join_tuples_inner(keys_a, keys_b, swap)
-        }
-        (_, _, 1, 1) => {
-            let keys_a = splitted_a
-                .iter()
-                .map(|ca| {
-                    ca.downcast_iter()
-                        .flat_map(|v| v.into_iter().map(|v| v.copied().as_u64()))
-                        .collect::<Vec<_>>()
-                })
-                .collect::<Vec<_>>();
-
-            let keys_b = splitted_b
-                .iter()
-                .map(|ca| {
-                    ca.downcast_iter()
-                        .flat_map(|v| v.into_iter().map(|v| v.copied().as_u64()))
-                        .collect::<Vec<_>>()
-                })
-                .collect::<Vec<_>>();
-            hash_join_tuples_inner(keys_a, keys_b, swap)
-        }
-        _ => {
-            let keys_a = splitted_a
-                .iter()
-                .map(|ca| ca.into_iter().map(|v| v.as_u64()).collect::<Vec<_>>())
-                .collect::<Vec<_>>();
-            let keys_b = splitted_b
-                .iter()
-                .map(|ca| ca.into_iter().map(|v| v.as_u64()).collect::<Vec<_>>())
-                .collect::<Vec<_>>();
-            hash_join_tuples_inner(keys_a, keys_b, swap)
-        }
-    }
-}
-
-fn num_group_join_left<T>(
-    left: &ChunkedArray<T>,
-    right: &ChunkedArray<T>,
-) -> Vec<(IdxSize, Option<IdxSize>)>
-where
-    T: PolarsIntegerType,
-    T::Native: Hash + Eq + Send + AsU64,
-    Option<T::Native>: AsU64,
-{
-    let n_threads = POOL.current_num_threads();
-    let splitted_a = split_ca(left, n_threads).unwrap();
-    let splitted_b = split_ca(right, n_threads).unwrap();
-    match (
-        left.null_count(),
-        right.null_count(),
-        left.chunks.len(),
-        right.chunks.len(),
-    ) {
-        (0, 0, 1, 1) => {
-            let keys_a = splitted_a
-                .iter()
-                .map(|ca| ca.cont_slice().unwrap())
-                .collect::<Vec<_>>();
-            let keys_b = splitted_b
-                .iter()
-                .map(|ca| ca.cont_slice().unwrap())
-                .collect::<Vec<_>>();
-            hash_join_tuples_left(keys_a, keys_b)
-        }
-        (0, 0, _, _) => {
-            let keys_a = splitted_a
-                .iter()
-                .map(|ca| ca.into_no_null_iter().collect_trusted::<Vec<_>>())
-                .collect::<Vec<_>>();
-            let keys_b = splitted_b
-                .iter()
-                .map(|ca| ca.into_no_null_iter().collect_trusted::<Vec<_>>())
-                .collect::<Vec<_>>();
-            hash_join_tuples_left(keys_a, keys_b)
-        }
-        (_, _, 1, 1) => {
-            let keys_a = splitted_a
-                .iter()
-                .map(|ca| {
-                    // we know that we only iterate over length == self.len()
-                    unsafe {
-                        ca.downcast_iter()
-                            .flat_map(|v| v.into_iter().map(|v| v.copied().as_u64()))
-                            .trust_my_length(ca.len())
-                            .collect_trusted::<Vec<_>>()
-                    }
-                })
-                .collect::<Vec<_>>();
-
-            let keys_b = splitted_b
-                .iter()
-                .map(|ca| {
-                    // we know that we only iterate over length == self.len()
-                    unsafe {
-                        ca.downcast_iter()
-                            .flat_map(|v| v.into_iter().map(|v| v.copied().as_u64()))
-                            .trust_my_length(ca.len())
-                            .collect_trusted::<Vec<_>>()
-                    }
-                })
-                .collect::<Vec<_>>();
-            hash_join_tuples_left(keys_a, keys_b)
-        }
-        _ => {
-            let keys_a = splitted_a
-                .iter()
-                .map(|ca| {
-                    ca.into_iter()
-                        .map(|v| v.as_u64())
-                        .collect_trusted::<Vec<_>>()
-                })
-                .collect::<Vec<_>>();
-            let keys_b = splitted_b
-                .iter()
-                .map(|ca| {
-                    ca.into_iter()
-                        .map(|v| v.as_u64())
-                        .collect_trusted::<Vec<_>>()
-                })
-                .collect::<Vec<_>>();
-            hash_join_tuples_left(keys_a, keys_b)
-        }
-    }
-}
-
-impl<T> HashJoin<T> for ChunkedArray<T>
-where
-    T: PolarsIntegerType + Sync,
-    T::Native: Eq + Hash + num::NumCast,
-{
-    fn hash_join_outer(&self, other: &ChunkedArray<T>) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        let (a, b, swap) = det_hash_prone_order!(self, other);
-
-        let n_partitions = set_partition_size();
-        let splitted_a = split_ca(a, n_partitions).unwrap();
-        let splitted_b = split_ca(b, n_partitions).unwrap();
-
-        match (a.null_count(), b.null_count()) {
-            (0, 0) => {
-                let iters_a = splitted_a
-                    .iter()
-                    .map(|ca| ca.into_no_null_iter())
-                    .collect::<Vec<_>>();
-                let iters_b = splitted_b
-                    .iter()
-                    .map(|ca| ca.into_no_null_iter())
-                    .collect::<Vec<_>>();
-                hash_join_tuples_outer(iters_a, iters_b, swap)
-            }
-            _ => {
-                let iters_a = splitted_a
-                    .iter()
-                    .map(|ca| ca.into_iter())
-                    .collect::<Vec<_>>();
-                let iters_b = splitted_b
-                    .iter()
-                    .map(|ca| ca.into_iter())
-                    .collect::<Vec<_>>();
-                hash_join_tuples_outer(iters_a, iters_b, swap)
-            }
-        }
-    }
-}
-
-impl HashJoin<BooleanType> for BooleanChunked {
-    fn hash_join_outer(&self, other: &BooleanChunked) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        let (a, b, swap) = det_hash_prone_order!(self, other);
-
-        let n_partitions = set_partition_size();
-        let splitted_a = split_ca(a, n_partitions).unwrap();
-        let splitted_b = split_ca(b, n_partitions).unwrap();
-
-        match (a.null_count(), b.null_count()) {
-            (0, 0) => {
-                let iters_a = splitted_a
-                    .iter()
-                    .map(|ca| ca.into_no_null_iter())
-                    .collect::<Vec<_>>();
-                let iters_b = splitted_b
-                    .iter()
-                    .map(|ca| ca.into_no_null_iter())
-                    .collect::<Vec<_>>();
-                hash_join_tuples_outer(iters_a, iters_b, swap)
-            }
-            _ => {
-                let iters_a = splitted_a
-                    .iter()
-                    .map(|ca| ca.into_iter())
-                    .collect::<Vec<_>>();
-                let iters_b = splitted_b
-                    .iter()
-                    .map(|ca| ca.into_iter())
-                    .collect::<Vec<_>>();
-                hash_join_tuples_outer(iters_a, iters_b, swap)
-            }
-        }
-    }
-}
-
-pub(crate) fn prepare_strs<'a>(
-    been_split: &'a [Utf8Chunked],
-    hb: &RandomState,
-) -> Vec<Vec<StrHash<'a>>> {
-    POOL.install(|| {
-        been_split
-            .par_iter()
-            .map(|ca| {
-                ca.into_iter()
-                    .map(|opt_s| {
-                        let mut state = hb.build_hasher();
-                        opt_s.hash(&mut state);
-                        let hash = state.finish();
-                        StrHash::new(opt_s, hash)
-                    })
-                    .collect::<Vec<_>>()
-            })
-            .collect()
-    })
-}
-
-impl HashJoin<Utf8Type> for Utf8Chunked {
-    fn hash_join_inner(&self, other: &Utf8Chunked) -> Vec<(IdxSize, IdxSize)> {
-        let n_threads = POOL.current_num_threads();
-
-        let (a, b, swap) = det_hash_prone_order!(self, other);
-
-        let hb = RandomState::default();
-        let splitted_a = split_ca(a, n_threads).unwrap();
-        let splitted_b = split_ca(b, n_threads).unwrap();
-
-        let str_hashes_a = prepare_strs(&splitted_a, &hb);
-        let str_hashes_b = prepare_strs(&splitted_b, &hb);
-        hash_join_tuples_inner(str_hashes_a, str_hashes_b, swap)
-    }
-
-    fn hash_join_left(&self, other: &Utf8Chunked) -> Vec<(IdxSize, Option<IdxSize>)> {
-        let n_threads = POOL.current_num_threads();
-
-        let hb = RandomState::default();
-        let splitted_a = split_ca(self, n_threads).unwrap();
-        let splitted_b = split_ca(other, n_threads).unwrap();
-
-        let str_hashes_a = prepare_strs(&splitted_a, &hb);
-        let str_hashes_b = prepare_strs(&splitted_b, &hb);
-        hash_join_tuples_left(str_hashes_a, str_hashes_b)
-    }
-
-    fn hash_join_outer(&self, other: &Utf8Chunked) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        let (a, b, swap) = det_hash_prone_order!(self, other);
-
-        let n_partitions = set_partition_size();
-        let splitted_a = split_ca(a, n_partitions).unwrap();
-        let splitted_b = split_ca(b, n_partitions).unwrap();
-
-        match (a.has_validity(), b.has_validity()) {
-            (false, false) => {
-                let iters_a = splitted_a
-                    .iter()
-                    .map(|ca| ca.into_no_null_iter())
-                    .collect::<Vec<_>>();
-                let iters_b = splitted_b
-                    .iter()
-                    .map(|ca| ca.into_no_null_iter())
-                    .collect::<Vec<_>>();
-                hash_join_tuples_outer(iters_a, iters_b, swap)
-            }
-            _ => {
-                let iters_a = splitted_a
-                    .iter()
-                    .map(|ca| ca.into_iter())
-                    .collect::<Vec<_>>();
-                let iters_b = splitted_b
-                    .iter()
-                    .map(|ca| ca.into_iter())
-                    .collect::<Vec<_>>();
-                hash_join_tuples_outer(iters_a, iters_b, swap)
-            }
-        }
-    }
 }
 
 pub trait ZipOuterJoinColumn {
