+++ polars/polars-core/src/series/implementations/categorical.rs
@@ -12,7 +12,7 @@ use crate::chunked_array::{
 };
 use crate::fmt::FmtList;
 use crate::frame::groupby::*;
-use crate::frame::hash_join::{check_categorical_src, HashJoin, ZipOuterJoinColumn};
+use crate::frame::hash_join::{check_categorical_src, ZipOuterJoinColumn};
 use crate::prelude::*;
 use crate::series::implementations::SeriesWrap;
 use ahash::RandomState;
@@ -107,11 +107,6 @@ impl private::PrivateSeries for SeriesWrap<CategoricalChunked> {
         })
     }
 
-    fn hash_join_outer(&self, other: &Series) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        self.0
-            .logical()
-            .hash_join_outer(other.categorical().unwrap().logical())
-    }
     fn zip_outer_join_column(
         &self,
         right_column: &Series,
