+++ polars/polars-core/src/series/implementations/dates_time.rs
@@ -134,10 +134,6 @@ macro_rules! impl_dyn_series {
                     .map(|s| s.$into_logical().into_series())
             }
 
-            fn hash_join_outer(&self, other: &Series) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-                let other = other.to_physical_repr().into_owned();
-                self.0.hash_join_outer(&other.as_ref().as_ref())
-            }
             fn zip_outer_join_column(
                 &self,
                 right_column: &Series,
