+++ polars/polars-core/src/series/implementations/datetime.rs
@@ -138,10 +138,6 @@ impl private::PrivateSeries for SeriesWrap<DatetimeChunked> {
                 .into_series()
         })
     }
-    fn hash_join_outer(&self, other: &Series) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        let other = other.to_physical_repr().into_owned();
-        self.0.hash_join_outer(other.as_ref().as_ref())
-    }
     fn zip_outer_join_column(
         &self,
         right_column: &Series,
