+++ polars/polars-core/src/series/implementations/duration.rs
@@ -140,10 +140,6 @@ impl private::PrivateSeries for SeriesWrap<DurationChunked> {
             .map(|s| s.into_duration(self.0.time_unit()).into_series())
     }
 
-    fn hash_join_outer(&self, other: &Series) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-        let other = other.to_physical_repr().into_owned();
-        self.0.hash_join_outer(other.as_ref().as_ref())
-    }
     fn zip_outer_join_column(
         &self,
         right_column: &Series,
