+++ polars/polars-core/src/series/series_trait.rs
@@ -204,9 +204,6 @@ pub(crate) mod private {
         fn agg_median(&self, _groups: &GroupsProxy) -> Option<Series> {
             None
         }
-        fn hash_join_outer(&self, _other: &Series) -> Vec<(Option<IdxSize>, Option<IdxSize>)> {
-            invalid_operation_panic!(self)
-        }
         fn zip_outer_join_column(
             &self,
             _right_column: &Series,
